+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0rc2 (unreleased)
 -------------------
 
+- Added default video as a link to a MP4 file and videos default view
+  [jeanferri]
+
 - Added a link to MP3 example file in /institucional/audio
   [jeanferri]